requirements.txt
discord==1.7.3
discord-pretty-help==1.3.4